+++ tests/testthat/test-methods.R
@@ -170,21 +170,21 @@ test_that("as.person method", {
   )
 })
 
-test_that("as person with another cff", {
+test_that("Errors on other as.person methods", {
   path <- system.file("examples/CITATION_complete.cff", package = "cffr")
   the_cff <- cff_read(path)
-  expect_s3_class(the_cff, "cff")
-  expect_identical(as.person(the_cff), person())
+  expect_s3_class(the_cff, "cff", exact = TRUE)
+  expect_snapshot(as.person(the_cff), error = TRUE)
 
   # identifiers
   key <- the_cff$identifiers
-  expect_s3_class(key, "cff")
-  expect_identical(as.person(key), person())
+  expect_s3_class(key, c("cff_ref_list", "cff"), exact = TRUE)
+  expect_snapshot(as.person(key), error = TRUE)
 
   # preferred
   key <- the_cff$`preferred-citation`
-  expect_s3_class(key, "cff")
-  expect_identical(as.person(key), person())
+  expect_s3_class(key, c("cff_ref", "cff"), exact = TRUE)
+  expect_snapshot(as.person(key), error = TRUE)
 })
 
 test_that("head and tail", {
@@ -250,6 +250,28 @@ test_that("toBibtex", {
 
   froml <- toBibtex(cff_read_bib_text(string))
   expect_equal(sum(names(froml) == "title"), 1)
+
+
+  # Persons
+
+  sev_auth <- as_cff_person(
+    "{The Big Bopper} and Frank Sinatra and Dean Martin and Davis, Jr., Sammy"
+  )
+
+  expect_length(sev_auth, 4)
+  expect_s3_class(sev_auth, "cff_pers_list")
+  expect_snapshot(toBibtex(sev_auth))
+
+
+  # Single person
+  single <- as_cff_person(person("A", "person", email = "a@b.d"))[[1]]
+  expect_s3_class(single, "cff_pers")
+  expect_snapshot(toBibtex(single))
+
+  # Single entity
+  single <- as_cff_person(person("{A and B co}", email = "a@b.d"))[[1]]
+  expect_s3_class(single, "cff_pers")
+  expect_snapshot(toBibtex(single))
 })